*: stub out networkhelpers
